requirements.txt
flask
google-generativeai
PyMuPDF
pdfplumber
reportlab
PyPDF2
python-dotenv
requests


